avm2: Move top-level functions and __ruffle__ stub functions to ActionScript
