--- core/src/avm2/globals/__ruffle__/logging.as
@@ -1,11 +1,0 @@
-package __ruffle__ {
-    public native function log_warn(...arguments);
-
-    public native function stub_method(class_name, method);
-
-    public native function stub_getter(class_name, method);
-
-    public native function stub_setter(class_name, method);
-
-    public native function stub_constructor(class_name, method);
-}
