+++ core/src/avm2/globals/__ruffle__/stubs.as
@@ -1,0 +1,9 @@
+package __ruffle__ {
+    public native function stub_method(... rest):void;
+
+    public native function stub_getter(... rest):void;
+
+    public native function stub_setter(... rest):void;
+
+    public native function stub_constructor(... rest):void;
+}
