+++ core/src/avm2/globals/stubs.as
@@ -16,4 +16,3 @@ include "String.as"
 include "int.as"
 include "uint.as"
 include "Vector.as"
-include "__ruffle__/logging.as"
