+++ core/src/avm2/globals/toplevel.rs
@@ -5,11 +5,10 @@ use ruffle_wstr::Units;
 use crate::avm2::activation::Activation;
 use crate::avm2::error::{uri_error, Error};
 use crate::avm2::object::Object;
+use crate::avm2::parameters::ParametersExt;
 use crate::avm2::value::Value;
 use crate::string::{AvmString, WStr, WString};
-use crate::stub::Stub;
 use ruffle_wstr::Integer;
-use std::borrow::Cow;
 use std::fmt::Write;
 
 pub fn trace<'gc>(
@@ -36,168 +35,24 @@ pub fn trace<'gc>(
     Ok(Value::Undefined)
 }
 
-pub fn log_warn<'gc>(
-    activation: &mut Activation<'_, 'gc>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    match args {
-        [] => tracing::warn!("(__ruffle__.log_warn called with no arg)"),
-        [arg] => {
-            let msg = arg.coerce_to_string(activation)?;
-            tracing::warn!("{}", &msg.to_utf8_lossy());
-        }
-        args => {
-            let strings = args
-                .iter()
-                .map(|a| a.coerce_to_string(activation))
-                .collect::<Result<Vec<_>, _>>()?;
-            let msg = crate::string::join(&strings, &WStr::from_units(b" "));
-            tracing::warn!("{}", &msg.to_utf8_lossy());
-        }
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn stub_method<'gc>(
-    activation: &mut Activation<'_, 'gc>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    match args {
-        [class, method] => {
-            let class = class.coerce_to_string(activation)?;
-            let method = method.coerce_to_string(activation)?;
-            activation
-                .context
-                .stub_tracker
-                .encounter(&Stub::Avm2Method {
-                    class: Cow::Owned(class.to_utf8_lossy().to_string()),
-                    method: Cow::Owned(method.to_utf8_lossy().to_string()),
-                    specifics: None,
-                });
-        }
-        [class, method, specifics] => {
-            let class = class.coerce_to_string(activation)?;
-            let method = method.coerce_to_string(activation)?;
-            let specifics = specifics.coerce_to_string(activation)?;
-            activation
-                .context
-                .stub_tracker
-                .encounter(&Stub::Avm2Method {
-                    class: Cow::Owned(class.to_utf8_lossy().to_string()),
-                    method: Cow::Owned(method.to_utf8_lossy().to_string()),
-                    specifics: Some(Cow::Owned(specifics.to_utf8_lossy().to_string())),
-                });
-        }
-        _ => tracing::warn!("(__ruffle__.stub_method called with wrong args)"),
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn stub_getter<'gc>(
-    activation: &mut Activation<'_, 'gc>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    match args {
-        [class, property] => {
-            let class = class.coerce_to_string(activation)?;
-            let property = property.coerce_to_string(activation)?;
-            activation
-                .context
-                .stub_tracker
-                .encounter(&Stub::Avm2Getter {
-                    class: Cow::Owned(class.to_utf8_lossy().to_string()),
-                    property: Cow::Owned(property.to_utf8_lossy().to_string()),
-                });
-        }
-        _ => tracing::warn!("(__ruffle__.stub_getter called with wrong args)"),
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn stub_setter<'gc>(
-    activation: &mut Activation<'_, 'gc>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    match args {
-        [class, property] => {
-            let class = class.coerce_to_string(activation)?;
-            let property = property.coerce_to_string(activation)?;
-            activation
-                .context
-                .stub_tracker
-                .encounter(&Stub::Avm2Setter {
-                    class: Cow::Owned(class.to_utf8_lossy().to_string()),
-                    property: Cow::Owned(property.to_utf8_lossy().to_string()),
-                });
-        }
-        _ => tracing::warn!("(__ruffle__.stub_setter called with wrong args)"),
-    }
-
-    Ok(Value::Undefined)
-}
-
-pub fn stub_constructor<'gc>(
-    activation: &mut Activation<'_, 'gc>,
-    _this: Object<'gc>,
-    args: &[Value<'gc>],
-) -> Result<Value<'gc>, Error<'gc>> {
-    match args {
-        [class] => {
-            let class = class.coerce_to_string(activation)?;
-            activation
-                .context
-                .stub_tracker
-                .encounter(&Stub::Avm2Constructor {
-                    class: Cow::Owned(class.to_utf8_lossy().to_string()),
-                    specifics: None,
-                });
-        }
-        [class, specifics] => {
-            let class = class.coerce_to_string(activation)?;
-            let specifics = specifics.coerce_to_string(activation)?;
-            activation
-                .context
-                .stub_tracker
-                .encounter(&Stub::Avm2Constructor {
-                    class: Cow::Owned(class.to_utf8_lossy().to_string()),
-                    specifics: Some(Cow::Owned(specifics.to_utf8_lossy().to_string())),
-                });
-        }
-        _ => tracing::warn!("(__ruffle__.stub_constructor called with wrong args)"),
-    }
-
-    Ok(Value::Undefined)
-}
-
 pub fn is_finite<'gc>(
     activation: &mut Activation<'_, 'gc>,
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(val) = args.get(0) {
-        Ok(val.coerce_to_number(activation)?.is_finite().into())
-    } else {
-        Ok(false.into())
-    }
+    let val = args.get_f64(activation, 0)?;
+
+    Ok(val.is_finite().into())
 }
 
-pub fn is_nan<'gc>(
+pub fn is_na_n<'gc>(
     activation: &mut Activation<'_, 'gc>,
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(val) = args.get(0) {
-        Ok(val.coerce_to_number(activation)?.is_nan().into())
-    } else {
-        Ok(true.into())
-    }
+    let val = args.get_f64(activation, 0)?;
+
+    Ok(val.is_nan().into())
 }
 
 pub fn parse_int<'gc>(
@@ -205,15 +60,8 @@ pub fn parse_int<'gc>(
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    let string = match args.get(0).unwrap_or(&Value::Undefined) {
-        Value::Undefined => "null".into(),
-        value => value.coerce_to_string(activation)?,
-    };
-
-    let radix = match args.get(1) {
-        Some(value) => value.coerce_to_i32(activation)?,
-        None => 0,
-    };
+    let string = args.get_string(activation, 0)?;
+    let radix = args.get_i32(activation, 1)?;
 
     let result = crate::avm2::value::string_to_int(&string, radix, false);
     Ok(result.into())
@@ -224,15 +72,14 @@ pub fn parse_float<'gc>(
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    if let Some(value) = args.get(0) {
-        let string = value.coerce_to_string(activation)?;
-        let swf_version = activation.context.swf.version();
-        if let Some(result) = crate::avm2::value::string_to_f64(&string, swf_version, false) {
-            return Ok(result.into());
-        }
-    }
+    let string = args.get_string(activation, 0)?;
+    let swf_version = activation.context.swf.version();
 
-    Ok(f64::NAN.into())
+    if let Some(result) = crate::avm2::value::string_to_f64(&string, swf_version, false) {
+        Ok(result.into())
+    } else {
+        Ok(f64::NAN.into())
+    }
 }
 
 pub fn is_xml_name<'gc>(
@@ -240,7 +87,7 @@ pub fn is_xml_name<'gc>(
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    let name = args.get(0).unwrap_or(&Value::Undefined);
+    let name = args.get_value(0);
     if matches!(name, Value::Undefined | Value::Null) {
         return Ok(false.into());
     }
@@ -255,11 +102,7 @@ pub fn escape<'gc>(
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    let value = match args.first() {
-        None => return Ok("undefined".into()),
-        Some(Value::Undefined) => return Ok("null".into()),
-        Some(value) => value,
-    };
+    let value = args.get_string(activation, 0)?;
 
     let mut output = WString::new();
 
@@ -267,7 +110,7 @@ pub fn escape<'gc>(
     let not_converted =
         WStr::from_units(b"0123456789abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ@-_.*+/");
 
-    for x in value.coerce_to_string(activation)?.iter() {
+    for x in value.iter() {
         if not_converted.contains(x) {
             output.push(x);
         } else {
@@ -288,11 +131,7 @@ pub fn unescape<'gc>(
     _this: Object<'gc>,
     args: &[Value<'gc>],
 ) -> Result<Value<'gc>, Error<'gc>> {
-    let value = match args.first() {
-        None => return Ok("undefined".into()),
-        Some(Value::Undefined) => return Ok("null".into()),
-        Some(value) => value.coerce_to_string(activation)?,
-    };
+    let value = args.get_string(activation, 0)?;
 
     let mut output = WString::new();
     let mut index = 0;
@@ -361,15 +200,9 @@ fn encode_utf8_with_exclusions<'gc>(
     args: &[Value<'gc>],
     not_converted: &str,
 ) -> Result<Value<'gc>, Error<'gc>> {
-    let value = match args.first() {
-        None => return Ok("undefined".into()),
-        Some(Value::Undefined) => return Ok("null".into()),
-        Some(value) => value,
-    };
-
+    let input = args.get_string(activation, 0)?;
     let mut output = String::new();
 
-    let input = value.coerce_to_string(activation)?;
     let input_string = match input.units() {
         // Latin-1 values map directly to unicode codepoints,
         // so we can directly convert to a `char`
@@ -434,11 +267,7 @@ fn decode<'gc>(
     reserved_set: &str,
     func_name: &str,
 ) -> Result<Value<'gc>, Error<'gc>> {
-    let value = match args.first() {
-        None => return Ok("undefined".into()),
-        Some(Value::Undefined) => return Ok("null".into()),
-        Some(value) => value.coerce_to_string(activation)?,
-    };
+    let value = args.get_string(activation, 0)?;
 
     let mut output = WString::new();
     let mut chars = value.chars();
